+++ commands/client/saucenao.py
@@ -6,18 +6,25 @@ from library.ConsoleSelect import *
 from library.SauceSelect import *
 
 from pysaucenao import *
+from private.config import signature
 
 class SauceNow(commands.Cog):
     def __init__(self, client):
         self.client = client
 
+    # SAUCENAO GROUP
+    @commands.group()
+    async def saucenao(self, ctx):
+        if ctx.invoked_subcommand is None:
+            embed = discord.Embed (description = 'That is not a visible subcommand.', color = signature)
+            await ctx.reply(embed = embed)
+            
     # PIXNAO - METHOD
-    @commands.command(help = 'Delivers the sauce by image. [Pixiv]')
-    async def pixnao(self, ctx, image: str):
-        # INITIAL API INITALIZATION
+    @saucenao.command(help = 'Delivers the sauce by image. [Pixiv]', aliases = ['pix'])
+    async def pixiv(self, ctx, image: str):
         sauce = SauceNao(api_key = 'ad5dd123e3b1b2ccdd38575f2736287b0aea1e26')
         results = await sauce.from_url(url = image)
-        # INSTANCE CHECK
+        
         if isinstance(results[0], PixivSource):
             message: discord.Message = await ctx.reply('I found a match... <a:checkmark:903852585360949289> gimmie me a second... <a:cleo:902025889179631637>')
             await asyncio.sleep(1)
@@ -35,6 +42,7 @@ class SauceNow(commands.Cog):
             embed.add_field (name = 'Result Count', value = f'{len(results)}')
             embed.add_field (name = 'Index ID', value = f'{results[0].index_id}')
             embed.timestamp = message.created_at
+            
             await message.delete()
             await ctx.reply(embed = embed)
             print(f'[{Time.timeCST()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE: SAUCENAO.PY - LOG: SauceNao [Pixiv] was utilized in #{ctx.channel}! \n[Raw Data: {results[0].url}]')
@@ -44,12 +52,11 @@ class SauceNow(commands.Cog):
             await ctx.reply(embed = embed)
     
     # BOORUNAO - METHOD
-    @commands.command(help = 'Delivers the sauce by image. [Gelbooru/Danbooru]')
-    async def boorunao(self, ctx, image: str):
-        # INITIAL API INITALIZATION
+    @saucenao.command(help = 'Delivers the sauce by image. [Gelbooru/Danbooru]')
+    async def booru(self, ctx, image: str):
         sauce = SauceNao(api_key = 'ad5dd123e3b1b2ccdd38575f2736287b0aea1e26')
         results = await sauce.from_url(url = image); len(results)
-        # INSTANCE CHECK
+
         if isinstance(results[0], BooruSource):
             message: discord.Message = await ctx.reply('I found a match... <a:checkmark:903852585360949289> gimmie me a second... <a:cleo:902025889179631637>')
             await asyncio.sleep(1)
@@ -69,6 +76,7 @@ class SauceNow(commands.Cog):
             embed.set_footer (
                 text = f'Source URL: {results[0].source_url} Index: {results[0].index}'
             )
+            
             await message.delete()
             await ctx.reply(embed = embed)
             print(f'[{Time.timeCST()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE: SAUCENAO.PY - LOG: SauceNao [Gelbooru/Danbooru] was utilized in #{ctx.channel}!')