+++ docs/changelog.md
@@ -162,7 +162,7 @@
 
 - Event Handiling Changes, and listeners to fix other stuff.
 
-## [1.3.0] - 2021-10-31 - 2021-11-**
+## [1.3.0] - 2021-10-31 - 2021-11-14
 
 - `spotify` command replacing the `track` command.
 
@@ -179,3 +179,9 @@
 - Cache folder to hold all my images that are not assets. They are temp files and will be refreshed routinely.
 
 - Changed the way time works so that way it timestamps from other peoples time to prevent confusion.
+
+- Akaneko was refractored and given way more commands.
+
+- New `Neko (library)` and many neko commands incoming.
+
+- Group command features.